requirements.txt
alembic==1.3.3
asgiref==3.2.3
Babel==2.8.0
blinker==1.4
bp==0.0.4
Click==7.0
Django==3.0.2
django-appconf==1.0.3
dominate==2.4.0
elasticsearch==7.6.0
fanstatic==1.1
Flask==1.1.1
Flask-Babel==1.0.0
Flask-Bootstrap==3.3.7.1
Flask-Cors==3.0.8
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-Migrate==2.5.2
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-WTF==0.14.3
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.3
js.jquery==3.3.1
Mako==1.1.1
MarkupSafe==1.1.1
numpy==1.18.1
opencv-python==4.2.0.32
Pillow==7.0.0
Pillow-PIL==0.1.dev0
psycopg2==2.8.5
PyJWT==1.7.1
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytz==2019.3
shutilwhich==1.1.0
six==1.14.0
SQLAlchemy==1.3.13
sqlparse==0.3.0
static-jquery==3.2.1.0
urllib3==1.25.8
visitor==0.1.3
WebOb==1.8.6
Werkzeug==0.16.0
WTForms==2.2.1
